Moved ChatsFragment.kt and ChannelsFragment.kt to fragments.home_screen.chats_and_channels package
