+++ app/src/main/res/layout/fragment_channels.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/chat_bg"
-    tools:context=".ui.fragment.ChannelsFragment">
+    tools:context=".ui.fragments.home_screen.chats_and_channels.ChannelsFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
